requirements.txt
annotated-types==0.7.0
anthropic==0.7.7
anyio==3.7.1
asgiref==3.9.1
backoff==2.2.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.5
defusedxml==0.7.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fpdf2==2.8.3
fsspec==2025.7.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kubernetes==33.1.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.99.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
overrides==7.7.0
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.3.0
posthog==6.3.3
protobuf==6.31.1
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pypdf==3.17.0
pypdfium2==4.30.0
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
